requirements.txt
requests
pygetwindow
pywin32
pywinauto
python-dotenv